move download function to helper
